examples/requirements.txt
aioboto3
aiobotocore
aiofiles
aiohappyeyeballs
aiohttp
aioitertools
aiosignal
annotated-doc
annotated-types
anyio
argcomplete
ascii_colors
asgiref
asttokens
attrs
audioop-lts
autobahn
Automat
beautifulsoup4
boto3
botocore
cachetools
cbor2
certifi
cffi
chardet
charset-normalizer
click
colorama
colorclass
compressed-rtf
configparser
constantly
cryptography
daphne
decorator
distro
dnspython
docx2txt
easygui
ebcdic
email-validator
executing
extract-msg
fastapi
fastapi-cli
fastapi-cloud-cli
fastar
filelock
frozenlist
fsspec
future
google-auth
google-genai
gunicorn
h11
h2
hf-xet
hpack
httpcore
httptools
httpx
huggingface-hub
Hypercorn
hyperframe
hyperlink
idna
IMAPClient
Incremental
ipython
ipython_pygments_lexers
itsdangerous
jedi
Jinja2
jiter
jmespath
json_repair
jsonpickle
lark
lightrag-hku
lxml
markdown-it-py
MarkupSafe
matplotlib-inline
mdurl
mpmath
msgpack
msoffcrypto-tool
multidict
nano-vectordb
neo4j
networkx
numpy
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-cufile-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-cusparselt-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvshmem-cu12
nvidia-nvtx-cu12
olefile
oletools
ollama
openai
oracledb
orjson
packaging
pandas
parso
pcodedmp
pdfminer.six
pexpect
pillow
pip
pipmaster
priority
prompt_toolkit
propcache
ptyprocess
pure_eval
py-ubjson
pyasn1
pyasn1_modules
pycparser
pycryptodome
pydantic
pydantic_core
pydantic-extra-types
pydantic-settings
Pygments
PyMySQL
pyOpenSSL
pyparsing
pypinyin
python-dateutil
python-dotenv
python-multipart
python-pptx
pytz
pyvis
PyYAML
red-black-tree-mod
regex
requests
rich
rich-toolkit
rignore
rsa
RTFDE
s3transfer
safetensors
sentry-sdk
service-identity
setuptools
shellingham
six
sniffio
sortedcontainers
soupsieve
SpeechRecognition
stack-data
standard-aifc
standard-chunk
starlette
sympy
tenacity
tiktoken
tokenizers
torch
torchaudio
torchvision
tqdm
traitlets
transformers
triton
Twisted
txaio
typer
typer-slim
typing_extensions
typing-inspection
tzdata
tzlocal
ujson
urllib3
uvicorn
uvloop
uWSGI
watchfiles
wcwidth
websockets
wrapt
wsproto
xlrd
xlsxwriter
yarl
zope.interface